refactor: [ANDROAPP-5116] move bottomsheet to ui
