+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2022-12-21 12:58","gitSha":"6aae602c3"}
+{"buildTime":"2023-01-11 12:46","gitSha":"c972b456a"}
