+++ app/src/main/java/org/dhis2/usescases/eventsWithoutRegistration/eventCapture/EventCaptureActivity.java
@@ -30,11 +30,11 @@ import org.dhis2.commons.sync.ConflictType;
 import org.dhis2.commons.dialogs.AlertBottomDialog;
 import org.dhis2.commons.dialogs.CustomDialog;
 import org.dhis2.commons.dialogs.DialogClickListener;
-import org.dhis2.commons.dialogs.bottomsheet.BottomSheetDialog;
+import org.dhis2.ui.dialogs.bottomsheet.BottomSheetDialog;
 import org.dhis2.commons.popupmenu.AppMenuHelper;
 import org.dhis2.databinding.ActivityEventCaptureBinding;
-import org.dhis2.commons.dialogs.bottomsheet.BottomSheetDialogUiModel;
-import org.dhis2.commons.dialogs.bottomsheet.DialogButtonStyle;
+import org.dhis2.ui.dialogs.bottomsheet.BottomSheetDialogUiModel;
+import org.dhis2.ui.dialogs.bottomsheet.DialogButtonStyle;
 import org.dhis2.ui.ThemeManager;
 import org.dhis2.usescases.eventsWithoutRegistration.eventCapture.eventCaptureFragment.OnEditionListener;
 import org.dhis2.usescases.eventsWithoutRegistration.eventCapture.model.EventCompletionDialog;
