+++ commons/src/main/res/values/styles.xml
@@ -85,20 +85,12 @@
         <item name="android:background">@android:color/transparent</item>
     </style>
 
-    <style name="CustomBottomSheetDialogTheme" parent="Theme.Design.Light.BottomSheetDialog">
-        <item name="bottomSheetStyle">@style/CustomBottomSheetStyle</item>
-    </style>
-
     <style name="BreakTheGlassBottomSheetDialogTheme" parent="ThemeOverlay.MaterialComponents.BottomSheetDialog">
         <item name="bottomSheetStyle">@style/CustomBottomSheetStyle</item>
         <item name="android:windowIsFloating">false</item>
         <item name="android:windowSoftInputMode">adjustResize</item>
         <item name="colorControlActivated">?colorPrimary</item>
         <item name="colorControlHighlight">?colorPrimary</item>
-    </style>
-
-    <style name="CustomBottomSheetStyle" parent="Widget.Design.BottomSheet.Modal">
-        <item name="android:background">@android:color/transparent</item>
     </style>
 
     <style name="filterPeriodRadioButton" parent="Widget.AppCompat.CompoundButton.RadioButton">
