+++ ui-components/src/main/java/org/dhis2/ui/theme/Color.kt
@@ -7,3 +7,8 @@ val contrastLight = Color(0xE6FFFFFF)
 
 val programColorDark = Color(0xFF00BCD4)
 val programColorLight = Color(0xFF84FFFF)
+
+val textPrimary = Color(0xDE000000)
+val textSecondary = Color(0x8A000000)
+val warningColor = Color(0xFFFF9800)
+val colorPrimary = Color(0xFF2C98F0)
