+++ ui-components/src/main/res/drawable/ic_alert.xml
