+++ ui-components/src/main/res/drawable/ic_error_outline.xml
