+++ ui-components/src/main/res/drawable/ic_event_status_complete.xml
