+++ ui-components/src/main/res/drawable/ic_saved_check.xml
